* (bug 3393) Blank user names still possible (Michael Jard <kfaust@gmail.com>) * (bug 3363) Download percentage overflow (Martin Doucha <next_ghost@quick.cz>) * (bug 3390) MSVC project (Julian Priestley <juzley@gmail.com>) * For OS X and MinGW ports, don't -I code/SDL when USE_LOCAL_HEADERS is 0

+++ code/q3_ui/ui_connect.c
@@ -94,7 +94,7 @@ static void UI_DisplayDownloadInfo( const char *downloadName ) {
 	UI_DrawProportionalString( 8, 224, xferText, style, color_white );
 
 	if (downloadSize > 0) {
-		s = va( "%s (%d%%)", downloadName, downloadCount * 100 / downloadSize );
+		s = va( "%s (%d%%)", downloadName, (int)( (float)downloadCount * 100.0f / downloadSize ) );
 	} else {
 		s = downloadName;
 	}
